pyproject.toml
[project]
name = "imp_bot"
version = "1.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.2",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "aiohttp>=3.12.14",
    "urllib3>=2.5.0",
    "certifi>=2024.7.4",
    "urllib3>=2.5.0",
    "requests>=2.32.4"
]
